Delete extra file and move feemarket registration
